==> 1-lyric-downloading.ipynb <==
from bs4 import BeautifulSoup

import json
import re

import requests
import numpy as np
import pandas as pd

genres = {}

## Rap 

def get_artist(url): 
    url = url
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')

    ranker = []
    for artist in soup.findAll('meta', {'itemprop':'name'}):
        ranker.append(artist['content'].lower().strip())

    return ranker 

rap_url = "https://www.ranker.com/crowdranked-list/the-greatest-rappers-of-all-time"
rap_url2 = 'https://www.ranker.com/list/best-hip-hop-artists-in-2018/ranker-music'
rap_url3 = 'https://www.ranker.com/list/best-rappers-2019/ranker-music'
rap_url5 = "https://www.ranker.com/crowdranked-list/best-_90s-rappers?ref=collections&l=855723&collectionId=88"
rap_url4 = "https://www.ranker.com/list/best-rappers-2020/kareem-gantt?ref=collections&l=2750895&collectionId=88"
rap_urls = [rap_url, rap_url2, rap_url3, rap_url4, rap_url5]
rappers = [] 
for url in rap_urls: 
    rappers += get_artist(url)

remove = ['ranker', "the greatest rappers of all time",'the best rappers of 2019, ranked', 
          "the best rappers of 2018, ranked", "the best rappers of 2020, ranked", 
         "the dopest rappers of the '90s"]
rappers = [x for x in rappers if x not in remove] 
rapper_list = list(set(rappers))
print(rapper_list)
print(len(rapper_list))
genres['rap'] = rapper_list

## Rock, Modern Rock, Indie Rock  --> Rock

rock_url = 'https://www.ranker.com/crowdranked-list/the-best-rock-bands-of-all-time'
rock_url2 = "https://www.ranker.com/list/modern-rock-bands-and-musicians/reference"
rock_url3 = "https://www.ranker.com/list/indie-bands-and-artists/reference"

rock_urls = [rock_url, rock_url2, rock_url3]
rockers = [] 
for url in rock_urls: 
    rockers += get_artist(url)

remove = ['ranker', 'the best modern rock bands/artists', 'the best rock bands of all time', 
          'the best indie bands & artists', 'the best indie bands & artists']
rockers = [x for x in rockers if x not in remove] 
rocker_list = list(set(rockers))
print(rocker_list)
print(len(rocker_list))
genres['rock'] = rocker_list

## Metal 

url = 'https://www.thetoptens.com/top-heavy-metal-bands/'
response = requests.get(url)
soup = BeautifulSoup(response.text, 'html.parser')

# Find all commenters
commenters = []
for commenter in soup.findAll('p'):
    commenter_username = commenter.find('b')
    if commenter_username:
        commenters.append(commenter_username.get_text().lower().strip())

# Toptens (Metal)
metal_toptens = []
for artist in soup.findAll('b')[1:]:
    artist_name = artist.get_text().lower().strip()
    if artist_name not in commenters:
        if artist_name[0:3] != '1. ':
            metal_toptens.append(artist_name)

metal_url = "https://www.ranker.com/crowdranked-list/best-heavy-metal-bands-that-i-know?ref=collections&l=336573&collectionId=1547"
metal_toptens += get_artist(metal_url)
remove = ['ranker', 'the greatest heavy metal bands of all time']
metal_toptens = [x for x in metal_toptens if x not in remove] 
print(metal_toptens)
print(len(metal_toptens))
genres['metal'] = metal_toptens


## Country 

ctry_url = "https://www.ranker.com/list/top-country-artists-of-all-time/samantha-dillinger"
ctry_url2 = 'https://www.ranker.com/list/best-new-country-artists/ranker-music'

ctry_urls = [ctry_url, ctry_url2]
country = [] 
for url in ctry_urls: 
    country += get_artist(url)

remove = ['ranker', 'the top country artists of all time']
country = [x for x in country if x not in remove] 
country_list = list(set(country))
print(country_list)
print(len(country_list))
genres['country'] = country_list

## Jazz 

jazz_url = "https://www.ranker.com/list/best-jazz-bands-and-artists/ranker-music?ref=collections&l=101462&collectionId=1428"
smooth_jazz = "https://www.ranker.com/list/smooth-jazz-bands-and-musicians/ranker-music"
female_jazz = "https://www.ranker.com/list/best-female-jazz-singers/rockboy?ref=collections&l=101377&collectionId=1461"

jazz_urls = [jazz_url, smooth_jazz, female_jazz]
jazz = [] 
for url in jazz_urls: 
    jazz += get_artist(url)
    
remove = ['ranker', 'the very best jazz musicians & artists', 
          'make your voice heard: who are the best female jazz singers?', 
         'the best smooth jazz bands and artists of all time']
jazz = [x for x in jazz if x not in remove] 
jazz_list = list(set(jazz))
print(jazz_list)
print(len(jazz_list))
genres['jazz'] = jazz_list

## Folk 

folk_url = "https://www.ranker.com/list/folk-music-bands-and-musicians/ranker-music"
folk = get_artist(folk_url)
remove = ['ranker']
folk = [x for x in folk if x not in remove] 
folk_list = list(set(folk))
print(folk_list)
print(len(folk_list))
genres['folk'] = folk_list

## Pop 

# Billboard (Pop)
url = 'https://www.billboard.com/charts/greatest-of-all-time-pop-songs-artists'
response = requests.get(url)
soup = BeautifulSoup(response.text, 'html.parser')

pop_billboard = []
for artist in soup.findAll('a', {'href':re.compile('/music/')}):
    artist = artist.get_text().replace('\n', '')
    pop_billboard.append(artist.lower().strip())

pop = list(set(pop_billboard))
print(pop)
print(len(pop))
genres['pop'] = list(set(pop_billboard))

## R&B 

rb_url = "https://www.ranker.com/list/best-rhythm-and-blues-bands-and-artists/ranker-music?ref=collections&l=1055563&collectionId=2872"
rb_url1 = "https://www.ranker.com/list/best-rhythm-and-blues-artists-2019/music-lover"
rb_url2 = "https://www.ranker.com/list/best-r-and-b-artists-and-bands/ranker-music?ref=collections&l=997413&collectionId=2872"

rb_urls = [rb_url, rb_url1, rb_url2]
rb = [] 
for url in rb_urls:
    rb += get_artist(rb_url)

remove = ['ranker', 'the greatest r&b artists of all time']
rb = [x for x in rb if x not in remove] 
rb_list = list(set(rb))
print(rb_list)
print(len(rb_list))
genres['rb'] = rb_list

## Soul 

soul_url = "https://www.ranker.com/list/soul-music-bands-and-musicians/reference?ref=collections&l=337012&collectionId=1909"
soul = [] 
soul = get_artist(soul_url)
remove = ['ranker']
soul = [x for x in soul if x not in remove] 
soul_list = list(set(soul))
print(soul_list)
print(len(soul_list))
genres['soul'] = soul_list


# Pulling Lyrics 

print(list(genres.keys()))

client_secret = "PyIOc0fpU1kTrUgGmxfWZJTzPXxo4nwJY68im8ijCNgC1X23A-HpB4N4ojME83sGeS7qEh0hBTKksUTs1Tn8Tg"
client_id = "UxVTuW-k7r1aP_9-7LBhfCCHJAdkXhu-tQslckEDV5_P9ChNAXmRtRVwyhrxOc4k"
access_token = "h7LnkbwF2kidbhXTUAA9HexEkCFMlnt05D8Zs2Y8H9FTEBvl1tra1fceYNtnAyhN"

import lyricsgenius as genius
import json 
import time 
api = genius.Genius("h7LnkbwF2kidbhXTUAA9HexEkCFMlnt05D8Zs2Y8H9FTEBvl1tra1fceYNtnAyhN")
api.verbose = False # Turn off status messages
api.remove_section_headers = True # Remove section headers (e.g. [Chorus]) from lyrics when searching
api.skip_non_songs = False # Include hits thought to be non-songs (e.g. track lists)
api.excluded_terms = ["(Remix)", "(Live)"] # Exclude songs with these words in their title

rap_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
rock_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
country_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])


metal_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])


def get_lyrics(genre, artist, num_songs): 
    filename = str(artist) + ".json"
    while True: 
        try: 
            artist = api.search_artist(artist, max_songs=num_songs) 
            break 
        except: 
            pass 
    
    if artist is None: 
        return 
    rows = [] 
    if artist.songs is not None: 
        print("# of songs:" + str(len(artist.songs)))
        for i in range(len(artist.songs)): 
            rows.append((str(artist.songs[i].artist), 
                         "rap", str(artist.songs[i].title ), 
                         str(artist.songs[i].lyrics)))

    artist_df = pd.DataFrame(rows, columns =['artist', 'genre', 'title', 'lyrics'])
    # setting it as global 
    global rap_df, rock_df, country_df, jazz_df, folk_df, rb_df, metal_df, pop_df, soul_df 
    if genre == "rap": 
        rap_df = rap_df.append(artist_df)
    if genre == "rock": 
        rock_df = rock_df.append(artist_df)
    if genre == "country":
        country_df = country_df.append(artist_df)
    if genre == "jazz":
        jazz_df = jazz_df.append(artist_df)
    if genre == "folk":
        folk_df = folk_df.append(artist_df)
    if genre == "rb": 
        rb_df = rb_df.append(artist_df)
    if genre == "metal":
        metal_df = metal_df.append(artist_df)
    if genre == "pop":
        pop_df = pop_df.append(artist_df)
    if genre == "soul":
        soul_df = soul_df.append(artist_df)
    print("Saved " +str(artist))
    return 

rappers = genres["rap"]
print(rappers)

def fetch_all_lyrics(genre): 
    print("GENRE: " + str(genre))
    artists = genres[genre]
    print("there are " + str(len(artists)) + " artists")
    for art in artists: 
        get_lyrics(genre, art, 200)
    
#fetch_all_lyrics("rap")
#rap_df.to_csv("rappers.csv")

# # rap, rock, metal, country, jazz, folk, pop, r&b, soul
# fetch_all_lyrics("rock")
# rock_df.to_csv("rockers.csv")

fetch_all_lyrics("metal")
metal_df.to_csv("metal.csv")

country_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
fetch_all_lyrics("country")
country_df.to_csv("country.csv")

# jazz_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
# folk_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
# rb_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
# metal_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
# pop_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])


# fetch_all_lyrics("jazz")
# jazz_df.to_csv("jazz.csv")

# fetch_all_lyrics("folk")
# folk_df.to_csv("folk.csv")

# fetch_all_lyrics("pop")
# pop_df.to_csv("pop.csv")

# fetch_all_lyrics("rb")
# rb_df.to_csv("rb.csv")

soul_df = pd.DataFrame(columns=['artist', 'genre', 'title', 'lyrics'])
fetch_all_lyrics("soul")
soul_df.to_csv("soul.csv")

# other urls: 
# http://jdaytn.com/posts/download-blink-182-data/
# https://github.com/johnwmillr/LyricsGenius
# https://towardsdatascience.com/does-country-music-drink-more-than-other-genres-a21db901940b
