==> 2-pre-processing.ipynb <==
import csv
import numpy as np 
import pandas as pd 

## Genius Lyrics Cleaning 

metal = pd.read_csv("genius_data/metal.csv")
rap = pd.read_csv("genius_data/rappers.csv")
rock = pd.read_csv("genius_data/rockers.csv")
jazz = pd.read_csv("genius_data/jazz.csv")
folk = pd.read_csv("genius_data/folk.csv")
pop = pd.read_csv("genius_data/pop.csv")
rb = pd.read_csv("genius_data/rb.csv")
soul = pd.read_csv("genius_data/soul.csv")

frames = [metal, rap, rock, jazz, folk, pop, rb, soul]
genius_df = pd.concat(frames)
genius_df.to_csv("raw-genius-data.csv")
del genius_df['Unnamed: 0']
genius_df.head(5)

def clean_genius(genius_df): 
    #take out observations that have NaN for ‘genre’ or ‘lyrics’
    subset = genius_df[['genre','lyrics']]
    genius_df.dropna(subset=['genre', 'lyrics'], inplace=True)
    
    #replace new line with space
    genius_df = genius_df.replace({'\n': ' '}, regex=True)
    
    #get word count
    genius_df['word_num'] = genius_df['lyrics'].str.split().str.len()
    
    #observe that there are lots of songs w/ 1 word
    genius_df.sort_values(by = "word_num").head(100)
    
    #remove entries where only 1 word in song
    genius_df['word_num'].astype('int32')
    genius_df = genius_df[genius_df.word_num != 1]
    
    #anything with rap genius comments is way to long and includes comments
    genius_df = genius_df[~genius_df['lyrics'].str.contains("RAP GENIUS")]
    
    #notice that 124/130 of the songs with "Lyrics are just some string saying how there are no lyrics"
    genius_df = genius_df[genius_df['word_num'] !=18]
    
    #anything less than 10 seems to be junk lyrics
    genius_df = genius_df[genius_df['word_num'] > 10]



clean_genius(genius_df)
genius_df.head(5)

## Metrolyrics Cleaning

metro = pd.read_csv("lyrics-metrolyrics.csv")

def clean_metro(lyrics):
    #replace new line with space
    lyrics = lyrics.replace({'\n': ' '}, regex=True)
    
    #get word count
    lyrics['word_num'] = lyrics['lyrics'].str.split().str.len()

    #Take out index
    lyrics = lyrics.drop(columns = ['index'])

    #See what values are present for years
    column_values = lyrics[["year"]].values.ravel()
    unique_values =  pd.unique(column_values)
                                    
    #remove the dates that do not make any sense
    bad_year = ['702', '112', '67']
    lyrics = lyrics[~lyrics['year'].isin(bad_year)]

    #See what values are present for genre
    column_values = lyrics[["genre"]].values.ravel()
    unique_values =  pd.unique(column_values)

    #get the counts of each 
    index = pd.Index(lyrics['genre'])
    index.value_counts()
                                    
    #Remove all instrumental songs
    lyrics = lyrics[~lyrics.lyrics.str.contains("instrumental", na=False)]
    lyrics = lyrics[~lyrics.lyrics.str.contains("INSTRUMENTAL", na=False)]
    lyrics = lyrics[~lyrics.lyrics.str.contains("[Instrumental]", na=False)]

clean_metro(metro)
del metro['index']
del metro['year']
metro.head(5)

metro.rename(columns = {"song": "title"}, inplace=True)
# reordering columns 
metro = metro[['artist', 'genre', 'title', 'lyrics']]
metro['title'] = metro['title'].str.replace('-',' ')
metro.head(5)

# Combining datasets into BIG

frames_1 = [metro, genius_df]
master_df = pd.concat(frames_1)

master_df.shape
master_df.to_csv("master-data.csv")

# Cleaning Non-English Lyrics out of Dataset

from langdetect import detect
import pandas as pd 
import numpy as np 

data = pd.read_csv("master-data.csv")
del data['Unnamed: 0']
print(data.head())

# Get rid of all rows whose lyrics column is NaN
data = data[data['lyrics'].notnull()]
non_strings = 0
bad_indices = []
for index, value in data['lyrics'].items():
    if type(value) != str:
        non_strings = non_strings + 1
        bad_indices.append(index)
# print(data[type(data['lyrics']) != str].shape )
print(non_strings)
print(data.shape)
print(len(bad_indices))

print(len(data['lyrics']))

def language_detector(string):
    global i
    try:
        res = detect(string)
    except:
        res = "undetectable"
    if i % 1000 == 0:
        print(i)
    i = i + 1
    return res

i = 0
data['language'] = data['lyrics'].apply(language_detector)
data = data[data['language'] == "en"]
data.to_csv("language-processed-data.csv")

# Turning into master CSV 

data['t-lyric'] = data['title'] + " @@@ " + master_df['lyrics']
data['t-lyric'] = data['t-lyric'].str.lower()

data.head()

stopChars = [',','(',')','.','-','[',']','"']
# preprocessing the corpus by converting all letters to lowercase, 
# replacing blank lines with blank string and removing special characters
def preprocessText(text):
#     text = text.replace('\n', ' ').replace('\t','')
    processedText = text.lower()
    for char in stopChars:
        processedText = processedText.replace(char,'')
    return processedText

# tokenization 
def corpusToList(corpus):
    corpusList = [w for w in corpus.split(' ')] 
    corpusList = [i for i in corpusList if i] #removing empty strings from list
    return corpusList

# Preprocess 
data['lyrics'] = data['lyrics'].apply(preprocessText)
data['t-lyric'] = data['t-lyric'].apply(preprocessText)
data.head(5)

# turning the italicized characters into regular 

data['t-lyric'] = data['t-lyric'].apply(corpusToList)
data.head(5)

data.to_csv("master-process-data.csv")
